refactor(test): Y/N 입력뷰 테스트 중복 제거
Y/N 입력에 대한 검증 테스트가 반복되어 중복을 제거하였습니다.
